spritesheet_generator/requirements.txt
google-genai>=0.1.0
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.10.0
